Move to folder `Composite`
